twilio-experiment/requirements-prod.txt
google-cloud-aiplatform>=1.36.0
protobuf>=4.23.4
boto3>=1.28.0
wave>=0.0.2
aiohttp
aiortc
elevenlabs
fastapi
gunicorn
numpy
mutagen
dotenv
google-cloud-texttospeech
google-cloud-speech
google-genai
webrtcvad
amazon-transcribe
azure-cognitiveservices-speech
azure-identity
openai
pydantic
flask
twilio
redis
